+++ ComPASS/distributed_system_mswells.py
@@ -58,7 +58,7 @@ class DistributedSystemMSWells:
 
     @property
     def global_system_size(self):
-        local_size = np.array(self.local_system_size, dtype=np.int)
+        local_size = np.array(self.local_system_size, dtype=np.intc)
         global_size = np.zeros(local_size.shape, dtype=local_size.dtype)
         mpi.COMM_WORLD.Allreduce(local_size, global_size, mpi.SUM)
         return global_size[0], global_size[1]  # convert to tuple
